+++ sample/src/commonMain/kotlin/com/kmpbits/sample/android/data/dto/UploadResponse.kt
@@ -1,0 +1,6 @@
+package com.kmpbits.sample.android.data.dto
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class UploadResponse(val ok: Boolean)